NullPointerException from PlayerEngineImpl.build() fix
Because of wrong implementation of mapping tracks to albums in playlist,
it might happen that there would be some null pointers in the playlist map.

Mapping has been changed that now, if there is no album returned from the Jamendo
service for given track, the related playlist entry has a fake empty album set
as its album. This also assures that there will be no NullPointerException's,
in methods that require the album in playlist entry to be set.